Move block subsidy functions
