+++ divi/src/SuperblockHelpers.h
@@ -1,6 +1,8 @@
 #ifndef SUPERBLOCK_HELPERS_H
 #define SUPERBLOCK_HELPERS_H
 #include <stdint.h>
+#include <amount.h>
+
 class CBlockRewards;
 class CChainParams;
 
@@ -17,6 +19,8 @@ int64_t GetTreasuryReward(const CBlockRewards &rewards);
 int64_t GetCharityReward(const CBlockRewards &rewards);
 int64_t GetLotteryReward(const CBlockRewards &rewards);
 
+CBlockRewards GetBlockSubsidity(int nHeight);
+CAmount GetFullBlockValue(int nHeight);
 
 class LotteryAndTreasuryBlockSubsidyIncentives
 {
